policy: Change L4Filter.Port to uint16
Change L4Filter.Port to uint16 to avoid unnecessary conversions.

+++ pkg/endpoint/policy.go
@@ -93,7 +93,7 @@ func (e *Endpoint) getNamedPortEgress(npMap types.NamedPortMultiMap, name string
 // and the resolved destination port and protocol numbers, if any.
 // Must be called with e.mutex held.
 func (e *Endpoint) proxyID(l4 *policy.L4Filter, listener string) (string, uint16, uint8) {
-	port := uint16(l4.Port)
+	port := l4.Port
 	protocol := uint8(l4.U8Proto)
 	// Calculate protocol if it is 0 (default) and
 	// is not "ANY" (that is, it was not calculated).
